+++ demo/src/index.js
@@ -1,12 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { ThemeProvider, createMuiTheme } from '@material-ui/core/styles';
+import { grey } from '@material-ui/core/colors';
 import './index.css';
 import App from './App';
 
 const darkTheme = createMuiTheme({
   palette: {
-    type: 'dark'
+    type: 'dark',
+    primary: grey,
+    secondary: grey
+  },
+  props: {
+    MuiButtonBase: {
+      disableRipple: true
+    }
   }
 });