+++ src/index.js
@@ -1,13 +1,25 @@
 import { VivViewerLayer, StaticImageLayer } from './layers';
-import VivViewer from './VivViewer';
-import { createTiffPyramid, createZarrPyramid, ZarrLoader, createOMETiffLoader } from './loaders';
+import { VivViewer, PictureInPictureViewer, SideBySideViewer } from './viewers';
+import { VivView, OverviewView, DetailView, SideBySideView } from './views';
+import {
+  createTiffPyramid,
+  createZarrLoader,
+  ZarrLoader,
+  createOMETiffLoader
+} from './loaders';
 
 export {
   VivViewerLayer,
   VivViewer,
+  VivView,
+  OverviewView,
+  PictureInPictureViewer,
+  SideBySideView,
+  SideBySideViewer,
+  DetailView,
   StaticImageLayer,
   ZarrLoader,
   createTiffPyramid,
-  createZarrPyramid,
-  createOMETiffLoader
+  createOMETiffLoader,
+  createZarrLoader
 };