+++ src/loaders/tiffPyramidLoader.js
@@ -38,14 +38,16 @@ export default class TiffPyramidLoader {
      * Here we check if a tile is within bounds before requesting tile data.
      */
 
-    if (!this._tileInBounds({ x, y, z })) return null;
+    if (!this._tileInBounds({ x, y, z })) {
+      return { data: null, width: this.tileSize, height: this.tileSize };
+    }
 
     const tileRequests = this.channelPyramids.map(channelPyramid => {
       const image = channelPyramid[z];
       return this._getChannel({ image, x, y });
     });
-    const tiles = await Promise.all(tileRequests);
-    return tiles;
+    const data = await Promise.all(tileRequests);
+    return { data, width: this.tileSize, height: this.tileSize };
   }
 
   async getRaster({ z }) {