+++ tests/views/index.js
@@ -1,0 +1,4 @@
+require('./VivView.spec');
+require('./DetailView.spec');
+require('./OverviewView.spec');
+require('./SideBySideView.spec');